UI > Folder refactoring and renaming
