+++ lib/Character_Wiki/Data/Character/Models/character_model.dart
