+++ lib/Character_Wiki/Data/Character/Repositories/character_repository_impl.dart
@@ -1,6 +1,6 @@
 // ignore_for_file: public_member_api_docs, sort_constructors_first
-import 'package:character_wiki/Character_Wiki/Data/Character/DataSource/character_remote_data_source.dart';
-import 'package:character_wiki/Character_Wiki/Domain/Character/Entity/character.dart';
+import 'package:character_wiki/Character_Wiki/Data/Character/DataSources/character_remote_data_source.dart';
+import 'package:character_wiki/Character_Wiki/Domain/Character/Entities/character.dart';
 import 'package:character_wiki/Character_Wiki/Domain/Character/Repostories/character_repository.dart';
 
 class CharacterRepositoryImpl implements CharacterRepository {
