+++ lib/Character_Wiki/Data/Episode/Repositories/episode_repository_impl.dart
@@ -1,6 +1,6 @@
 // ignore_for_file: public_member_api_docs, sort_constructors_first
-import 'package:character_wiki/Character_Wiki/Data/Episode/DataSource/episode_remote_data_source.dart';
-import 'package:character_wiki/Character_Wiki/Domain/Episode/Entity/epsiode.dart';
+import 'package:character_wiki/Character_Wiki/Data/Episode/DataSources/episode_remote_data_source.dart';
+import 'package:character_wiki/Character_Wiki/Domain/Episode/Entities/epsiode.dart';
 import 'package:character_wiki/Character_Wiki/Domain/Episode/Repositories/epsiode_repository.dart';
 
 class EpisodeRepositoryImpl implements EpsiodeRepository {
