+++ lib/Character_Wiki/Data/Location/Repositories/location_repository_impl.dart
@@ -1,6 +1,6 @@
 // ignore_for_file: public_member_api_docs, sort_constructors_first
-import 'package:character_wiki/Character_Wiki/Data/Location/DataSource/location_remote_data_souce.dart';
-import 'package:character_wiki/Character_Wiki/Domain/Location/Entity/location.dart';
+import 'package:character_wiki/Character_Wiki/Data/Location/DataSources/location_remote_data_souce.dart';
+import 'package:character_wiki/Character_Wiki/Domain/Location/Entities/location.dart';
 import 'package:character_wiki/Character_Wiki/Domain/Location/Repositories/location_repository.dart';
 
 class LocationRepositoryImpl implements LocationRepository {
