+++ lib/Character_Wiki/Domain/Character/Entities/character.dart
