+++ lib/Character_Wiki/Domain/Character/UseCase/get_characters.dart
@@ -1,5 +1,5 @@
 // ignore_for_file: public_member_api_docs, sort_constructors_first
-import 'package:character_wiki/Character_Wiki/Domain/Character/Entity/character.dart';
+import 'package:character_wiki/Character_Wiki/Domain/Character/Entities/character.dart';
 import 'package:character_wiki/Character_Wiki/Domain/Character/Repostories/character_repository.dart';
 
 class GetCharacters {
