+++ lib/Character_Wiki/Domain/Episode/Entities/epsiode.dart
