+++ lib/Character_Wiki/Domain/Episode/Repositories/epsiode_repository.dart
@@ -1,4 +1,4 @@
-import 'package:character_wiki/Character_Wiki/Domain/Episode/Entity/epsiode.dart';
+import 'package:character_wiki/Character_Wiki/Domain/Episode/Entities/epsiode.dart';
 
 abstract class EpsiodeRepository {
   Future<List<Episode>> getEpisodes ();
