+++ lib/Character_Wiki/Domain/Location/Entities/location.dart
