+++ lib/Character_Wiki/Domain/Location/Repositories/location_repository.dart
@@ -1,4 +1,4 @@
-import 'package:character_wiki/Character_Wiki/Domain/Location/Entity/location.dart';
+import 'package:character_wiki/Character_Wiki/Domain/Location/Entities/location.dart';
 
 abstract class LocationRepository {
   Future<List<Location>> getLocations ();
