+++ lib/Character_Wiki/Presentation/Bloc/Character/character_bloc.dart
@@ -1,6 +1,6 @@
 // ignore_for_file: public_member_api_docs, sort_constructors_first
 import 'package:bloc/bloc.dart';
-import 'package:character_wiki/Character_Wiki/Domain/Character/Entity/character.dart';
+import 'package:character_wiki/Character_Wiki/Domain/Character/Entities/character.dart';
 import 'package:character_wiki/Character_Wiki/Domain/Character/UseCase/get_characters.dart';
 import 'package:equatable/equatable.dart';
 part 'character_event.dart';
